Refactored code so that it meets OOP standards.
Moved message creation logic to the Send class. Removed unnecessary methods and variables from the Receive class. The Receive class now only has variables that both the ESP and MCP share when sending message to the CCP. The ESP and MCP classes have their own unique private variables which don't overlap and their according getters/setters. Included checking that the message is intended for our CCP and BR Number.  Completed response logic i.e what should the CCP send when it receives from the ESP or MCP and who should it send it to.