Comments and Minor Changes
Included comments for all refactoring changes made. Created a temporary connection status testing loop (currently commented out). Included sending a STAT message to the ESP in the heartbeatTimer. Created an actual and intended light colour variable in the ReceiveESP class so that when the CCP receives a STAT message from the ESP, it can cross check. If the colour doesn't match, the previous EXEC message is sent.